move overwintering functions to overwinter.py
